move pprof to main
